requirements-dev.txt
# PyYAML for dumpdata
PyYAML==5.1
ipdb==0.12
black==18.9b0
Werkzeug==0.14.1
django-debug-toolbar==2.2
django_extensions==2.1.6
flake8==3.8.3
reorder-python-imports==2.3.5
pre-commit==1.15.2
mypy==0.720
mypy-extensions==0.4.1
pytest-django==3.5.1
pytest-cov==2.8.1
pydot==1.4.1
pyparsing==2.4.6